[DEV-1671] reorganizing the S3 helpers
